Код из файла main.js разбит по модулям вспомогательных функций util.js и генерации данных data.js
